requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
astro-sdk-python[amazon,snowflake] >= 1.1.0
apache-airflow-providers-postgres
berserk
python-dotenv